+++ pixi/src/ui/PageExperience.js
@@ -30,20 +30,13 @@ import InputBar from './InputBar.js';
 
 import getRecommendations from '../utils/checkAggregation/recommendations.js';
 import getStatusId from '../utils/checkAggregation/statusBanner';
+import getIssueUrl from '../utils/issueUrl';
 
 const totalNumberOfChecks =
   AmpLinterCheck.getCheckCount() +
   PageExperienceCheck.getCheckCount() +
   MobileFriendlinessCheck.getCheckCount() +
   SafeBrowsingCheck.getCheckCount();
-const NO_DATA = 'n/a';
-
-function parseScore(metric) {
-  const {data, unit} = metric;
-  return `${(data.numericValue / unit.conversion).toFixed(unit.digits)} ${
-    unit.name
-  }`;
-}
 
 export default class PageExperience {
   constructor() {
@@ -107,7 +100,7 @@ export default class PageExperience {
 
     const recommendations = await recommendationsPromise;
     if (recommendations.length > 0) {
-      const issueUrl = await this.getIssueUrl(
+      const issueUrl = await getIssueUrl(
         pageUrl,
         pageExperiencePromise,
         linterPromise,
@@ -253,47 +246,6 @@ export default class PageExperience {
     }
 
     return data;
-  }
-
-  async getIssueUrl(
-    pageUrl,
-    pageExperiencePromise,
-    linterPromise,
-    mobileFriendlinessPromise,
-    safeBrowsingPromise
-  ) {
-    const [
-      {pageExperience},
-      linter,
-      mobileFriendliness,
-      safeBrowsing,
-    ] = await Promise.all([
-      pageExperiencePromise,
-      linterPromise,
-      mobileFriendlinessPromise,
-      safeBrowsingPromise,
-    ]);
-    const hasFieldData =
-      pageExperience !== undefined && pageExperience.source === 'fieldData';
-    const issueData = {
-      lcp: hasFieldData ? parseScore(pageExperience.fieldData.lcp) : NO_DATA,
-      fid: hasFieldData ? parseScore(pageExperience.fieldData.fid) : NO_DATA,
-      cls: hasFieldData ? parseScore(pageExperience.fieldData.cls) : NO_DATA,
-      labLcp: parseScore(pageExperience.labData.lcp),
-      tbt: parseScore(pageExperience.labData.tbt),
-      labCls: parseScore(pageExperience.labData.cls),
-      safeBrowsing: safeBrowsing.safeBrowsing || NO_DATA,
-      mobileFriendly: mobileFriendliness.mobileFriendly || NO_DATA,
-      url: pageUrl,
-      usedComponents: linter.components || NO_DATA,
-      usesHttps:
-        linter.usesHttps === undefined
-          ? NO_DATA
-          : linter.usesHttps === true
-          ? 'pass'
-          : 'fail',
-    };
-    return `https://github.com/ampproject/amphtml/issues/new?assignees=&labels=Type%3A+Page+experience&title=Pixi:+Poor+page+experience&body=URL%0A---%0A${issueData.url}%0A%0ADetails%0A---%0A%0A%7C%20Metric%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%20%7C%20Field%20data%20%7C%20Lab%20data%20%7C%0A%7C-----------------%7C------------%7C---------%7C%0A%7C%20LCP%20%7C%20${issueData.lcp}%20%7C%20${issueData.labLcp}%20%7C%0A%7C%20FID%20%20%7C%20${issueData.fid}%20%7C%20${issueData.tbt}%20%7C%0A%7C%20CLS%20%20%7C%20${issueData.cls}%20%7C%20${issueData.labCls}%20%7C%0A%7C%20HTTPS%20%20%7C%20${issueData.usesHttps}%20%7C%20---%20%7C%0A%7C%20Safe%20browsing%20%20%7C%20${issueData.safeBrowsing}%20%7C%20---%20%7C%0A%7C%20Mobile-friendliness%20%20%7C%20${issueData.mobileFriendly}%20%7C%20---%20%7C%0A%7C%20Intrusive%20Interstitials%20%7C%20%3Cpass%2Ffail%3E%20%7C%20---%20%7C%0A%0ANotes%0A---%0A%0AComponents%20in%20use%3A%20${issueData.usedComponents}%0A%0A%0A%3C%21--%0A%3CAdditional%20notes%3E%0A--%3E%0A%0A%2Fcc%20%40ampproject%2Fwg-performance%60`;
   }
 
   toggleLoading(force) {
